完成Material的CRUD，修正Account和Doctor的部分依赖关系

+++ src/main/java/com/qst/medical/mapper/AccountMapper.java
@@ -24,7 +24,7 @@ public interface AccountMapper {
     /*校验手机号存在*/
     int checkPhone(String phone);
 
-    int selectIdByPhone(String phone);
+    Long selectIdByPhone(String phone);
 
     /*更新账户信息*/
     int updateAccount(AccountEntity entity);
